+++ packages/controls/js/libs/icon-control/index.js
@@ -272,7 +272,7 @@ function IconControl({
 							<Tooltip
 								text={__('Remove Icon', 'blockera')}
 								style={{
-									'--tooltip-bg': '#e20000',
+									'--tooltip-bg': '#e20b0b',
 								}}
 								delay={300}
 							>