+++ packages/wordpress/php/RenderBlock/V1/Render.php
@@ -215,7 +215,7 @@ class Render {
 
 			// If custom css is set, add it to the block css.
 			if (! empty($attributes['blockeraCustomCSS']['value'])) {
-				$css .= preg_replace('/(\.|#)block/i', $unique_class_name, $attributes['blockeraCustomCSS']['value']);
+				$css .= preg_replace(['/(\.|#)block/i', '/&/i'], $unique_class_name, $attributes['blockeraCustomCSS']['value']);
 			}
 
             // Print css into inline style on "wp_head" action occur.
@@ -254,7 +254,7 @@ class Render {
 
 		// If custom css is set, add it to the block css.
 		if (! empty($attributes['blockeraCustomCSS']['value'])) {
-			$computed_css_rules .= preg_replace('/(\.|#)block/i', $unique_class_name, $attributes['blockeraCustomCSS']['value']);
+			$computed_css_rules .= preg_replace(['/(\.|#)block/i', '/&/i'], $unique_class_name, $attributes['blockeraCustomCSS']['value']);
 		}
 
         // Print css into inline style on "wp_head" action occur.