+++ packages/editor/CHANGELOG.md
@@ -8,6 +8,9 @@
 - Added support for third-party developers to extend and customize the style engine filters.
 - Made the application faster by identifying and fixing slow areas of the code 🚀.
 - Improved how styles are applied to blocks to ensure they follow web standards and work reliably across different browsers.
+- The inner block items design improved to make them different of the block states.
+- Block Settings Section: improve design and UX.
+- Add `&` support for targeting current block in custom CSS feature.
 
 ## 2.0.1 (2025-07-22)